Alpaca_Algo/dashboard-prod/requirements.txt
dash==2.17.1
dash-ag-grid==31.2.0
dash-bootstrap-components==1.5.0
plotly==5.18.0
pandas==2.2.0
numpy==1.26.3
yfinance==0.2.50
pandas-ta==0.3.14b0
gunicorn==21.2.0
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
stripe==5.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# ── Auto Blog ──────────────────────────────────────────────────────────────
openai>=1.30.0
requests>=2.31.0
schedule>=1.2.0
Pillow>=10.0.0
